fix(layout): adjust presentation and video handling for sidebarless clients
This commit introduces specific handling for clients without a sidebar
(i.e., those with 'userdata-bbb_hide_sidebar_navigation=true'), which do not
accommodate layouts where presentation and camera dock are placed at the bottom
of the sidebar content(presentation focus and video focus).

To address this, the layout now mimics the behavior used on landscape tablets:
the presentation and the camera dock are mutually exclusive and share the same
space in the media area. When the presentation is hidden, the camera dock is
shown in its place.