fix(user-list): uneven spacing on paginated user list
Fixes an issue where users were not evenly spaced on user list pages beyond
the first one (i.e., when the meeting has more than 50 users). The problem was
caused by an extra `<div>` leftover from the skeleton loading state, which was
not removed after the page was fully loaded.