feat: Implement user consent modal for moderator unmute (#23384)
* feat: Implement user consent modal for unmute.
New feature flag `modsRequestToUnmuteUsers`.

When this feature is enabled, a moderator's action to unmute a user is changed. Instead of unmuting the user directly, the system now renders a modal to the end-user.

This modal prompts the user to approve or deny the request to enable their microphone.
- If the user **accepts**, their microphone will be unmuted.
- If the user **denies**, the modal is closed, and the user remains muted.

**Critical Dependency:**
For the `modsRequestToUnmuteUsers` flag to have any effect, the `allowModsToUnmuteUsers` flag must be set to `true`. If `allowModsToUnmuteUsers` is `false`, moderators do not have the base permission to unmute, and this new request flow will not be triggered.

* fix: changed flag name and PR corrections

* fix: reset flag for user reconnections

* Update bigbluebutton-web/grails-app/conf/bigbluebutton.properties

Co-authored-by: Arthur B. Grossi <32987232+Arthurk12@users.noreply.github.com>

* Update docs/docs/administration/customize.md

Co-authored-by: Anton Georgiev <antobinary@users.noreply.github.com>

* Update docs/docs/data/create.tsx

Co-authored-by: Anton Georgiev <antobinary@users.noreply.github.com>

* Update docs/docs/administration/customize.md

Co-authored-by: Anton Georgiev <antobinary@users.noreply.github.com>

* fix: changed modal subtitle

---------

Co-authored-by: Anton Georgiev <antobinary@users.noreply.github.com>
Co-authored-by: Arthur B. Grossi <32987232+Arthurk12@users.noreply.github.com>