fix(layout): local push layout being ovewritten
Adds a check to prevent the push layout value of the first user to join
the conference from being ovewritten to false due to a false indication
that the push layout of the meeting has changed. That occurs because
there is no previous props at the very first render of the push layout
engine, so we ensure prevProps is there and not undefined.