reroute logic, apply fallback to non default too
Many applications seem to remember which last device was connected and change the media stream to them, even if they weren't the system default. This is an another sloppy approach to these even edgier situations, but now we treat reconnecting devices that weren't the default, but were disconnected and reconnected after some time.

Strings 'default' and 'communications' returned from enumerateDevices obfuscate tracking by deviceId, so they're hardcoded for now.