fix(layout): remove incorrect sidebar navigation collapse action
Removes an incorrect dispatch that attempted to close the navigation sidebar
when `show_participants_on_login` is false. This behavior was correct in 3.0,
where the user list was within the navigation sidebar. However, in 3.1,
the navigation sidebar is no longer collapsible (except on mobile),
and the user list has its own separate panel.