fix(layout): apply meetingLayout and hide_presentation_on_join properly
This commit removes an incorrect condition introduced in ebf4e27570fb,
which prevented `meetingLayout` from being applied when `pushLayout` was
false.

It also fixes an issue where `hide_presentation_on_join=true` was not
respected. The push layout engine wrongly replicated the default server
value (presentation open), because the current user data from the GraphQL
subscription was undefined. As a result, the user was misidentified as
not being the presenter on the first render.