fix(layout): add `setByUserId` to differentiate pushed layouts from initial values
Adds a `setByUserId` property to the layout table to distinguish between default
values and those explicitly pushed by the presenter. This distinction is necessary
to properly hide the presentation when userdata is passed.

Also adds a check to only trigger the presentation pile process when there is
an actual change in the presentation area content.